Acceso_a_Datos/PR9 - Acceso por Voz/Asistente_Miriam/requirements.txt
psycopg2-binary
SpeechRecognition
pyaudio # permite a python comunicarse con el hardware del micro
python-dotenv